review_app/views.py
from .serializers import ApplicantSerializer
from .models import Applicant

from rest_framework import status
from rest_framework import generics

class ApplicantListCreateView(generics.ListCreateAPIView):
    queryset = Applicant.objects.all()
    serializer_class = ApplicantSerializer

class ApplicantUpdateDeleteView(generics.RetrieveUpdateDestroyAPIView):
    queryset = Applicant.objects.all()
    serializer_class = ApplicantSerializer


