requirements.txt
aiohttp==3.7.4.post0
anticaptchaofficial==1.0.31
apturl==0.5.2
asgiref==3.4.1
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
bcrypt==3.1.7
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.0.7
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
discord.py==1.7.3
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.7
djangorestframework==3.13.1
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.1.0
fasteners==0.14.1
future==0.18.2
greenlet==1.1.2
h11==0.12.0
httplib2==0.14.0
idna==2.8
kazam==1.4.5
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.6.3
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
multidict==5.2.0
netifaces==0.10.4
numpy==1.21.3
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.9
outcome==1.1.0
pandas==1.3.4
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
py==1.10.0
pycairo==1.16.2
pycodestyle==2.7.0
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==21.0.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.26.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
selenium==4.0.0
simplejson==3.16.0
six==1.14.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
SQLAlchemy==1.4.26
sqlparse==0.4.2
systemd-python==234
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==3.10.0.2
ubuntu-advantage-tools==27.10
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.7
usb-creator==0.3.7
uvicorn==0.11.3
wadllib==1.3.3
websockets==8.1
wsproto==1.0.0
xkit==0.0.0
xlwt==1.3.0
yarl==1.7.0
